// src/types/restaurant.d.ts
export type Menu = {
  categories: Category[];
  options: Option[];
  optionGroups: OptionGroup[];
  products: Product[];
  allergens: Allergen[];
  discounts?: Discount[];
};

export type Category = {
  id: string;
  name: string;
  description?: string[];
  imageUrl?: string;
  productIds: string[];
  discountId?: string[];
};

export type Option = {
  id: string;
  name: string;
  minAmount: number;
  maxAmount: number;
  prices: Prices;
  isSoldOut?: boolean;
  isVegetarian?: boolean;
  isVegan?: boolean;
  alcoholVolume?: number;
  caffeineAmount?: number;
  allergenIds?: string[];
};

export type Prices = {
  onsite: number;
  delivery?: number;
  pickup?: number;
};

export type OptionGroup = {
  id: string;
  name: string;
  optionIds: string[];
  minChoices?: number;
  maxChoices?: number;
  isTypeMulti: boolean;
  isRequired?: boolean;
};

export type Product = {
  id: string;
  name: string;
  description?: string[];
  imageUrl?: string;
  variants: Variant[];
  allergenIds?: string[];
};

export type Variant = {
  id: string;
  name?: string;
  optionGroupIds?: string[];
  prices: Prices;
  isSoldOut?: boolean;
  isVegetarian?: boolean;
  isVegan?: boolean;
  alcoholVolume?: number;
  caffeineAmount?: number;
  discountId?: string;
};

export type Allergen = {
  id: string;
  description: string;
};

export type Discount = {
  id: string;
  name: string;
  description: string;
  daysOfWeek: number[];
  from: number;
  until: number;
  discount?: number;
  discountedOptionGroupId?: string[];
};